requirements.txt
python-telegram-bot>=20.7
openai>=1.12.0
pandas>=2.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0
